packages/agent-python/pyproject.toml
[project]
name = "voice-dev-agent"
version = "1.0.0"
description = "LiveKit voice agent with MCP support"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[mcp,anthropic,openai,silero,turn-detector]>=1.0.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = "livekit.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch.*"
ignore_missing_imports = true
